poetry and presentation order change
presentation change was to flow with the hero banner. Question I can answer is actual order. highest order bit in content, do we rearrange hero banner to read that way too?  eg hero banner is a, b, c where c is the big win, but in detail we are now c, b, a cause we want them to see it in the case of TL;DR viewers